move LoraTestChatPayload to tests/utils/payloads
